requirements.txt
# Core dependencies
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.19.4

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

# LLM integration
ollama==0.1.7
openai==1.3.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
